Some Authorization added

+++ ClothingStoreAPI/Services/ClothingStoreService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using ClothingStoreAPI.Authorization;
 using ClothingStoreAPI.Entities;
 using ClothingStoreAPI.Entities.DbContextConfigure;
 using ClothingStoreAPI.Exceptions;
@@ -6,6 +7,7 @@ using ClothingStoreAPI.Services.Interfaces;
 using ClothingStoreModels.Dtos;
 using ClothingStoreModels.Dtos.Dispaly;
 using ClothingStoreModels.Dtos.Update;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 
 namespace ClothingStoreAPI.Services
@@ -15,17 +17,24 @@ namespace ClothingStoreAPI.Services
         private readonly ClothingStoreDbContext dbContext;
         private readonly IMapper mapper;
         private readonly ILogger<ClothingStoreService> logger;
+        private readonly IAuthorizationService authorizationService;
+        private readonly IUserContextService userContextService;
 
-        public ClothingStoreService(ClothingStoreDbContext dbContext, IMapper mapper, ILogger<ClothingStoreService> logger)
+        public ClothingStoreService(ClothingStoreDbContext dbContext, IMapper mapper,
+            ILogger<ClothingStoreService> logger, IAuthorizationService authorizationService,
+            IUserContextService userContextService)
         {
             this.dbContext = dbContext;
             this.mapper = mapper;
             this.logger = logger;
+            this.authorizationService = authorizationService;
+            this.userContextService = userContextService;
         }
 
         public int Create(CreateClothingStoreDto dto)
         {
             var store = mapper.Map<ClothingStore>(dto);
+            store.CreatedById = userContextService.GetUserId;
             dbContext.ClothingStores.Add(store);
             dbContext.SaveChanges();
 
@@ -37,11 +46,21 @@ namespace ClothingStoreAPI.Services
             logger.LogError($"Clothing Store with id: {id} DELETE ACTION invoked");
 
             var store = this.GetStoreFromDb(id);
+
+            var authorizationResult = authorizationService
+                .AuthorizeAsync(userContextService.User, store,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("You don't have access to someone else's store");
+            }
+
             dbContext.ClothingStores.Remove(store);
             dbContext.SaveChanges();
         }
 
-        public IEnumerable<ClothingStoreDto> GetAll()
+        public IEnumerable<ClothingStoreDto> GetAll(string searchPhraze)
         {
             var stores = dbContext
                 .ClothingStores
@@ -49,6 +68,7 @@ namespace ClothingStoreAPI.Services
                 .Include(s => s.Owner)
                 .Include(s => s.StoreReviews)
                 .Include(s => s.Products)
+                .Where(s => searchPhraze != null && s.Name.ToLower().Contains(searchPhraze.ToLower()))
                 .ToList();
 
             if(stores is null)
@@ -70,6 +90,15 @@ namespace ClothingStoreAPI.Services
         public void Update(UpdateClothingStoreDto dto, int id)
         {
             var store = this.GetStoreFromDb(id);
+
+            var authorizationResult = authorizationService
+                .AuthorizeAsync(userContextService.User, store,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if(!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("You don't have access to someone else's store");
+            }
 
             store = mapper.Map(dto, store);
             dbContext.SaveChanges();
